moved the two scripts into a new seperate top level folder called `scripts` and updated some imports in these two files
