+++ can/__init__.py
@@ -34,7 +34,7 @@ from .io import SqliteWriter, SqliteReader
 from .util import set_logging_level
 
 from .message import Message
-from .bus import BusABC
+from .bus import BusABC, BusState
 from .thread_safe_bus import ThreadSafeBus
 from .notifier import Notifier
 from .interfaces import VALID_INTERFACES
